Add temp files to script test, general
