+++ tests/script/general_mode/aip_directory/CD001_Images/Thumbs.db
@@ -1,0 +1,1 @@
+Placeholder for Thumbs.db
