+++ tests/script/general_mode/aip_directory/FD001_Text/Spreadsheet/.Worksheet.csv
@@ -1,0 +1,12 @@
+Letter,Number,Date
+a,1,1/1/2020
+b,2,1/2/2020
+a,3,1/3/2020
+b,4,1/4/2020
+a,5,1/5/2020
+b,6,1/6/2020
+a,7,1/7/2020
+b,8,1/8/2020
+a,9,1/9/2020
+b,10,1/10/2020
+a,11,1/11/2020
